+++ src/decompress.c
@@ -12,15 +12,15 @@
  * 
  * @param[in] pc_input_data Input data to be decompressed
  * @param[in] u64_input_data_size Size of the input data
- * @param[in] pc_output_data Buffer to hold the decompressed output data
+ * @param[in] ppc_output_data Pointer to the buffer that will hold the decompressed output data
  * @param[in out] pu64_output_data_size Pointer to hold the size of the decompressed data
  * @return s32 SUCCESS_STATUS on success, error code otherwise 
  */
-static s32 s32_rle_decompress(const char *pc_input_data, const u64 u64_input_data_size, char *pc_output_data, u64 *pu64_output_data_size)
+static s32 s32_rle_decompress(const char *pc_input_data, const u64 u64_input_data_size, char **ppc_output_data, u64 *pu64_output_data_size)
 {
     s32 s32_ret_val = FAILURE_STATUS;
 
-    if (NULL == pc_input_data || NULL == pc_output_data || NULL == pu64_output_data_size)
+    if (NULL == pc_input_data || NULL == *ppc_output_data || NULL == pu64_output_data_size)
     {
         s32_ret_val = ERROR_NULL_POINTER;
     }
@@ -39,7 +39,7 @@ static s32 s32_rle_decompress(const char *pc_input_data, const u64 u64_input_dat
         u64 u64_write_idx = 0;
         u64 u64_needed_size = 0; // Variable to hold needed size for realloc checks
 
-        char *pc_output_data_shadow = pc_output_data; // Shadow pointer to free in case of realloc failure
+        char *pc_output_data_shadow = *ppc_output_data; // Shadow pointer to free in case of realloc failure
 
         for (u64 i = 0; i < u64_input_data_size; i++)
         {
@@ -57,9 +57,15 @@ static s32 s32_rle_decompress(const char *pc_input_data, const u64 u64_input_dat
                 {
                     non_digit_char = pc_input_data[i + 1];
                 }
-                else
+                else if ('\\' == pc_input_data[i + 1])
                 {
                     non_digit_char = '\\';
+                }
+                else
+                {
+                    LOG_ERROR("Invalid escape sequence in compressed data.");
+                    s32_ret_val = ERROR_INVALID_DATA;
+                    break;
                 }
                 i++; // Skip the next character as it's part of the escape sequence
             }
@@ -68,7 +74,7 @@ static s32 s32_rle_decompress(const char *pc_input_data, const u64 u64_input_dat
                 non_digit_char = pc_input_data[i];
             }
 
-            pc_output_data[u64_write_idx++] = non_digit_char;
+            (*ppc_output_data)[u64_write_idx++] = non_digit_char;
 
             i++;
             while ((pc_input_data[i] >= '0') && (pc_input_data[i] <= '9'))
@@ -83,16 +89,16 @@ static s32 s32_rle_decompress(const char *pc_input_data, const u64 u64_input_dat
 
             u64_needed_size = u64_write_idx + u64_char_cnt + 1; // +1 for the already written character
 
-            if ((u64_needed_size / 2) < UINT32_MAX)
+            if ((u64_needed_size + (*pu64_output_data_size) + DATA_CHUNK_SIZE_BYTES) < UINT64_MAX)
             {
                 if (u64_needed_size >= *pu64_output_data_size)
                 {
                     LOG("Reallocating memory for decompression buffer.");
 
                     *pu64_output_data_size += DATA_CHUNK_SIZE_BYTES;
-                    pc_output_data = (char *)realloc(pc_output_data, *pu64_output_data_size);
+                    *ppc_output_data = (char *)realloc(*ppc_output_data, *pu64_output_data_size);
 
-                    if (NULL == pc_output_data)
+                    if (NULL == *ppc_output_data)
                     {
                         LOG_ERROR("Error reallocating memory for decompression buffer: %s", strerror(errno));
                         s32_ret_val == ERROR_MEMORY_ALLOCATION_FAILED;
@@ -111,7 +117,7 @@ static s32 s32_rle_decompress(const char *pc_input_data, const u64 u64_input_dat
 
             for (u64 j = 1; j < u64_char_cnt; j++)
             {
-                pc_output_data[u64_write_idx++] = non_digit_char;
+                (*ppc_output_data)[u64_write_idx++] = non_digit_char;
             }
         }
 
@@ -125,8 +131,8 @@ static s32 s32_rle_decompress(const char *pc_input_data, const u64 u64_input_dat
 
             LOG("Reallocating decompression buffer to the actual decompressed size.");
 
-            pc_output_data = (char *)realloc(pc_output_data, *pu64_output_data_size);
-            if (NULL == pc_output_data)
+            *ppc_output_data = (char *)realloc(*ppc_output_data, *pu64_output_data_size);
+            if (NULL == *ppc_output_data)
             {
                 LOG_ERROR("Error reallocating memory to the actual decompressed size: %s", strerror(errno));
                 s32_ret_val = ERROR_MEMORY_ALLOCATION_FAILED;
@@ -184,16 +190,26 @@ s32 decompress(const char *input_file_name)
                 break;
             }
 
-            s32_ret_val = open_file(input_file_name, "r", &pf_in_file);
-            ERROR_BREAK(s32_ret_val);
+            if (true == check_file_exists(input_file_name))
+            {
+                s32_ret_val = open_file(input_file_name, "r", &pf_in_file);
+                ERROR_BREAK(s32_ret_val);
 
-            s32_ret_val = read_file(pf_in_file, &pc_raw_data_buff, &u64_raw_data_size);
-            ERROR_BREAK(s32_ret_val);
+                s32_ret_val = read_file(pf_in_file, &pc_raw_data_buff, &u64_raw_data_size);
+                ERROR_BREAK(s32_ret_val);
 
-            s32_ret_val = close_file(&pf_in_file);
-            ERROR_BREAK(s32_ret_val);
+                s32_ret_val = close_file(&pf_in_file);
+                ERROR_BREAK(s32_ret_val);
+            }
+            else
+            {
+                LOG_ERROR("Input file does not exist: %s", input_file_name);
+                s32_ret_val = ERROR_FILE_NOT_FOUND;
+                break;
+            }
 
             u64_decompressed_size = u64_raw_data_size / 2;  // best scenario of size
+
             pc_decompressed_buff = (char *)malloc(u64_decompressed_size);
 
             if (NULL == pc_decompressed_buff)
@@ -203,7 +219,7 @@ s32 decompress(const char *input_file_name)
                 break;
             }
 
-            s32_ret_val = s32_rle_decompress(pc_raw_data_buff, u64_raw_data_size, pc_decompressed_buff, &u64_decompressed_size);
+            s32_ret_val = s32_rle_decompress(pc_raw_data_buff, u64_raw_data_size, &pc_decompressed_buff, &u64_decompressed_size);
             ERROR_BREAK(s32_ret_val);
 
             s32_ret_val = create_output_file(input_file_name, "txt", &pc_out_file_path);
@@ -225,6 +241,9 @@ s32 decompress(const char *input_file_name)
         // Clean-up
         if (SUCCESS_STATUS != s32_ret_val)
         {
+            // Save the error code before clean-up
+            s32 s32_err = s32_ret_val;
+
             LOG_ERROR("Exit decompression loop with error code: %d", s32_ret_val);
 
             if (NULL != pf_in_file)
@@ -237,10 +256,16 @@ s32 decompress(const char *input_file_name)
                 s32_ret_val = close_file(&pf_out_file);
             }
 
-            if (true == check_file_exists(pc_out_file_path))
+            if (NULL != pc_out_file_path)
             {
-                s32_ret_val = delete_file(pc_out_file_path);
+                if (true == check_file_exists(pc_out_file_path))
+                {
+                    s32_ret_val = delete_file(pc_out_file_path);
+                }
             }
+
+            // Restore the original error code
+            s32_ret_val = s32_err;
         }
 
         // Free allocated memory
